% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuikR-geo-data.R
\name{geo_data}
\alias{geo_data}
\title{Get Geographic Data from TUIK}
\usage{
geo_data(
  variable_no = NULL,
  variable_level = NULL,
  variable_source = NULL,
  variable_period = NULL,
  variable_recnum = NULL
)
}
\arguments{
\item{variable_no}{Data Series Number}

\item{variable_level}{NUTS Level (2, 3, or 4)}

\item{variable_source}{Data Series Source ("medas" or "ilGostergeleri")}

\item{variable_period}{Data Series Period ("yillik" or "aylik")}

\item{variable_recnum}{Data Series Record Number (3, 5, or 24)}
}
\value{
A data tibble
}
\description{
Get Geographic Data from TUIK
}
\examples{
\dontrun{
# Get metadata for all available variables
geo_data()

# Get data for a specific variable at NUTS-2 level
geo_data(
  variable_level = 2,
  variable_no = "SNM-GK160951-O33303",
  variable_source = "medas",
  variable_period = "yillik",
  variable_recnum = 5
)
}

}
