% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuikr-package.R
\docType{package}
\name{tuikr-package}
\alias{tuikr}
\alias{tuikr-package}
\title{tuikr: Download Data File And Database Urls from TUIK}
\description{
Download Data File And Database Urls from TUIK (Turkish Statistical Institute). Provides functions to extract statistical tables, databases, and geographic data from the TUIK web portal and geographic statistics portal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/tuikr}
  \item \url{https://eremrah.com/tuikr/}
  \item Report bugs at \url{https://github.com/emraher/tuikr/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
